requirements.txt
accelerate==0.22.0
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.1.1
annotated-types==0.5.0
anyio==3.5.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
async-timeout==4.0.3
attrs==22.1.0
backcall==0.2.0
backoff==2.2.1
bcrypt==4.0.1
beautifulsoup4==4.12.2
bitsandbytes==0.41.1
bleach==4.1.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
chroma-hnswlib==0.7.3
chromadb==0.4.10
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comm==0.1.2
contourpy==1.1.0
cycler==0.11.0
dataclasses-json==0.5.14
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
einops==0.6.1
entrypoints==0.4
executing==0.8.3
Faker==19.3.1
fastapi==0.99.1
fastjsonschema==2.16.2
favicon==0.7.0
filelock==3.12.3
flatbuffers==23.5.26
fonttools==4.42.1
frozenlist==1.4.0
fsspec==2023.9.0
gitdb==4.0.10
GitPython==3.1.34
greenlet==2.0.2
h11==0.14.0
htbuilder==0.6.1
httptools==0.6.0
huggingface-hub==0.16.4
humanfriendly==10.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.1
ipykernel==6.25.0
ipython==8.12.2
ipython-genutils==0.2.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.17.3
jupyter_client==7.4.9
jupyter_core==5.3.0
jupyter-server==1.23.4
jupyterlab-pygments==0.1.2
kiwisolver==1.4.5
langchain==0.0.281
langsmith==0.0.33
lxml==4.9.2
Markdown==3.4.4
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.1
marshmallow==3.20.1
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==0.8.4
monotonic==1.6
more-itertools==10.1.0
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nbclassic==0.5.5
nbclient==0.5.13
nbconvert==6.5.4
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
notebook==6.5.4
notebook_shim==0.2.2
numexpr==2.8.5
numpy==1.25.2
onnxruntime==1.15.1
openai==0.28.0
overrides==7.4.0
packaging==23.1
pandas==2.1.0
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
pip==23.2.1
platformdirs==3.10.0
posthog==3.0.2
prometheus-client==0.14.1
prompt-toolkit==3.0.36
protobuf==4.24.2
psutil==5.9.0
pulsar-client==3.3.0
pure-eval==0.2.2
pyarrow==13.0.0
pycparser==2.21
pydantic==1.10.12
pydantic_core==2.6.3
pydeck==0.8.0
Pygments==2.15.1
pymdown-extensions==10.3
Pympler==1.0.1
pyparsing==3.0.9
pypdf==3.15.5
PyPika==0.48.9
pyreadline3==3.4.1
pyrsistent==0.18.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pytz-deprecation-shim==0.1.0.post0
pywin32==305.1
pywinpty==2.0.10
PyYAML==6.0.1
pyzmq==23.2.0
regex==2023.8.8
requests==2.31.0
rich==13.5.2
safetensors==0.3.3
scikit-learn==1.3.0
scipy==1.11.2
Send2Trash==1.8.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
setuptools==68.0.0
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
soupsieve==2.4
SQLAlchemy==2.0.20
st-annotated-text==4.0.0
stack-data==0.2.0
starlette==0.27.0
streamlit==1.26.0
streamlit-camera-input-live==0.2.0
streamlit-card==0.0.61
streamlit-embedcode==0.1.2
streamlit-extras==0.3.0
streamlit-faker==0.0.2
streamlit-image-coordinates==0.1.6
streamlit-keyup==0.2.0
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==1.0.2
sympy==1.12
tenacity==8.2.3
terminado==0.17.1
threadpoolctl==3.2.0
tiktoken==0.4.0
tinycss2==1.2.1
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tornado==6.3.2
tqdm==4.66.1
traitlets==5.7.1
transformers==4.33.0
typing_extensions==4.7.1
typing-inspect==0.9.0
tzdata==2023.3
tzlocal==4.3.1
urllib3==2.0.4
uvicorn==0.23.2
validators==0.22.0
watchdog==3.0.0
watchfiles==0.20.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.58.0
websockets==11.0.3
wheel==0.38.4
yarl==1.9.2
zipp==3.16.2
